pac1-py/pyproject.toml
[project]
name = "bitgn-pac1-py"
version = "0.1.0"
description = "Runnable Python sample for the BitGN PAC1 benchmark"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bitgn-api-connectrpc-python==0.9.0.1.20260323122825+4916788a0b02",
    "bitgn-api-protocolbuffers-python==34.1.0.1.20260323122825+4916788a0b02",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.uv]
# AICODE-NOTE: `harness_core/scripts/sdk-python.sh` rewrites the Buf SDK pins
# in this file after `buf push`; keep this project flat so the sample stays
# runnable without packaging ceremony.
# AICODE-NOTE: PAC1 depends on the public PCM runtime modules generated under
# `bitgn.vm.pcm_*`; if local pins predate that schema publication, rerun
# `make sdk-python` from `harness_core` after pushing the latest API digest.
package = false
